Basic Paxos functionality
Prompts the user to enter a transaction, which is then added to a transaction queue. After a timer expires, Paxos is initiated using the transaction queue as the value we are trying to agree on. Paxos initiation consists of spawning a leader() thread, which then spawns threads to communicate with each node and reach consensus. Each node has acceptor() threads waiting for to be contacted by a leader, which then start accept() threads to handle leader election and consensus. Thus, every node can have up to 3 (will be 5 later) instances of Paxos running at once, 1 as a leader and 2 as an acceptor.

What's missing:
- Depth updates and "learner" functionality. Depth is not currently used in any meaningful way
- Quorums. Right now each node communicates with the entire network.
- Crash and network failures.
- Commands to print state of node/network.